More refactoring 	- Create new folders to sort the files 	- Move more code to new files
