+++ src/main/java/ShopifyProj/Controller/CustomerController.java
@@ -1,5 +1,7 @@
-package ShopifyProj;
+package ShopifyProj.Controller;
 
+import ShopifyProj.Model.Customer;
+import ShopifyProj.Repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,7 +13,8 @@ public class CustomerController {
     private CustomerRepository customerRepository;
 
     @PostMapping("/createCustomerAccount")
-    public @ResponseBody Customer createCustomerAccount(@ModelAttribute Customer newCustomer, Model model) {
+    public @ResponseBody
+    Customer createCustomerAccount(@ModelAttribute Customer newCustomer, Model model) {
         customerRepository.save(newCustomer);
         return newCustomer;
     }
