+++ src/main/java/ShopifyProj/Controller/MerchantController.java
@@ -1,5 +1,7 @@
-package ShopifyProj;
+package ShopifyProj.Controller;
 
+import ShopifyProj.Model.Merchant;
+import ShopifyProj.Repository.MerchantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
