--- src/main/java/ShopifyProj/CustomerRepository.java
@@ -1,8 +1,0 @@
-package ShopifyProj;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
-@RepositoryRestResource(collectionResourceRel = "merchant", path = "customer")
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
-}
