+++ src/main/java/ShopifyProj/Model/Shop.java
@@ -1,4 +1,4 @@
-package ShopifyProj;
+package ShopifyProj.Model;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
