+++ src/main/java/ShopifyProj/Repository/ImageRepository.java
@@ -1,9 +1,10 @@
-package ShopifyProj;
+package ShopifyProj.Repository;
 
+import ShopifyProj.Model.Image;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "images", path = "images")
-public interface ImageRepository extends CrudRepository<Images, Integer> {
+public interface ImageRepository extends CrudRepository<Image, Integer> {
 
 }
