+++ src/main/java/ShopifyProj/Repository/MerchantRepository.java
@@ -1,0 +1,9 @@
+package ShopifyProj.Repository;
+
+import ShopifyProj.Model.Merchant;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+@RepositoryRestResource(collectionResourceRel = "merchant", path="merchant")
+public interface MerchantRepository extends CrudRepository<Merchant, Integer> {
+}
