+++ src/main/java/ShopifyProj/Repository/ShopRepository.java
@@ -1,5 +1,6 @@
-package ShopifyProj;
+package ShopifyProj.Repository;
 
+import ShopifyProj.Model.Shop;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
