+++ src/main/java/ShopifyProj/Repository/TagRepository.java
@@ -1,5 +1,6 @@
-package ShopifyProj;
+package ShopifyProj.Repository;
 
+import ShopifyProj.Model.Tag;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
