+++ src/main/resources/static/js/SearchController.js
