+++ src/main/resources/templates/AddShopPage.html
@@ -11,7 +11,7 @@
 <h1>Add Shop Page</h1>
 
 <h3>Your Current Shops:</h3>
-<form action="#" th:action="@{/YourShopPage}">
+<form action="#" th:action="@{/goToShopMerchantView}">
     <table id="shopIdTable">
         <tr>
             <th>Shop ID</th>
@@ -26,7 +26,7 @@
                 <button type="submit" name="shopId" th:value="${shop.id}">Go to your Shop</button>
             </td>
             <td>
-                <input type="button" value="View as Customer" th:onclick="|location.href='/goToShop?shopId=${shop.id}'|"/>
+                <input type="button" value="View as Customer" th:onclick="|location.href='/goToShopCustomerView?shopId=${shop.id}'|"/>
             </td>
         </tr>
     </table>
@@ -45,7 +45,7 @@
     <input type="submit" value="Go To Home Page"/>
 </form>
 
-<script type="text/javascript" th:src="@{/js/addShopController.js}"></script>
+<script type="text/javascript" th:src="@{/js/AddShopController.js}"></script>
 
 </body>
 </html>
