+++ src/main/resources/templates/CustomerShopPage.html
