+++ src/main/resources/templates/SearchPage.html
@@ -22,7 +22,7 @@
         <input type="submit" value="Go To Home Page"/>
     </form>
 
-<script type="text/javascript" th:src="@{/js/searchController.js}"></script>
+<script type="text/javascript" th:src="@{/js/SearchController.js}"></script>
 
 </body>
 </html>
