+++ src/test/java/WebApplicationTest.java
@@ -1,5 +1,6 @@
 package ShopifyProj;
 
+import ShopifyProj.Repository.ShopRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
